+++ .credo.exs
@@ -159,7 +159,7 @@
         #
         {Credo.Check.Readability.StrictModuleLayout, priority: :normal},
         {Credo.Check.Consistency.MultiAliasImportRequireUse, false},
-        {Credo.Check.Consistency.UnusedVariableNames, []},
+        {Credo.Check.Consistency.UnusedVariableNames, force: :meaningful},
         {Credo.Check.Design.DuplicatedCode, false},
         {Credo.Check.Readability.AliasAs, false},
         {Credo.Check.Readability.MultiAlias, false},